Removed REST config

+++ application/core/MY_Model.php
@@ -21,10 +21,27 @@ class MY_Model extends CI_Model
     protected $_table;
 
     /**
+     * Database conn object; will use default connection
+     * unless overridden
+     */
+    protected $_db;
+
+    /**
      * This model's default primary key or unique identifier.
      * Used by the get(), update() and delete() functions.
      */
     protected $primary_key = 'id';
+
+    /**
+     * Support for soft deletes and this model's 'deleted' key
+     */
+    protected $soft_delete = TRUE;
+    protected $soft_delete_key = 'deleted';
+    protected $_temporary_with_deleted = FALSE;
+    /**
+     * Excluded table attributes for join statements
+     */
+    protected $excluded_attr = array('id','created_at','updated_at','deleted','location_id');
 
     /**
      * The various callbacks available to the model. Each are
@@ -38,6 +55,17 @@ class MY_Model extends CI_Model
     protected $after_get = array();
     protected $before_delete = array();
     protected $after_delete = array();
+
+    protected $callback_parameters = array();
+
+    /**
+     * Relationship arrays. Use flat strings for defaults or string
+     * => array to customise the class name and primary key
+     */
+    protected $belongs_to = array();
+    protected $has_many = array();
+
+    protected $_with = array();
 
     /**
      * An array of validation rules. This needs to be the same format
@@ -72,7 +100,11 @@ class MY_Model extends CI_Model
 
         $this->load->helper('inflector');
 
+        $this->_set_database();
+
         $this->_fetch_table();
+
+        $this->soft_delete_key = $this->_table.'.deleted';
 
         $this->_temporary_return_type = $this->return_type;
     }
@@ -86,14 +118,19 @@ class MY_Model extends CI_Model
      */
     public function get($primary_value)
     {
-        $this->_run_before_callbacks('get');
+        $this->trigger('before_get');
 
-        $row = $this->db->where($this->primary_key, $primary_value)
+        if ($this->soft_delete && $this->_temporary_with_deleted !== TRUE)
+        {
+            $this->db->where($this->soft_delete_key, FALSE);
+        }
+
+        $row = $this->db->where($this->_table. '.' .$this->primary_key, $primary_value)
                         ->get($this->_table)
                         ->{$this->_return_type()}();
         $this->_temporary_return_type = $this->return_type;
 
-        $this->_run_after_callbacks('get', array( $row ));
+        $row = $this->trigger('after_get', $row);
 
         return $row;
     }
@@ -107,12 +144,18 @@ class MY_Model extends CI_Model
         $where = func_get_args();
         $this->_set_where($where);
 
-        $this->_run_before_callbacks('get');
+        if ($this->soft_delete && $this->_temporary_with_deleted !== TRUE)
+        {
+            $this->db->where($this->soft_delete_key, FALSE);
+        }
+
+        $this->trigger('before_get');
+
         $row = $this->db->get($this->_table)
                         ->{$this->_return_type()}();
         $this->_temporary_return_type = $this->return_type;
 
-        $this->_run_after_callbacks('get', array( $row ));
+        $row = $this->trigger('after_get', $row);
 
         return $row;
     }
@@ -122,6 +165,11 @@ class MY_Model extends CI_Model
      */
     public function get_many($values)
     {
+        if ($this->soft_delete && $this->_temporary_with_deleted !== TRUE)
+        {
+            $this->db->where($this->soft_delete_key, FALSE);
+        }
+
         $this->db->where_in($this->primary_key, $values);
 
         return $this->get_all();
@@ -135,6 +183,11 @@ class MY_Model extends CI_Model
         $where = func_get_args();
         $this->_set_where($where);
 
+        if ($this->soft_delete && $this->_temporary_with_deleted !== TRUE)
+        {
+            $this->db->where($this->soft_delete_key, FALSE);
+        }
+
         return $this->get_all();
     }
 
@@ -144,7 +197,12 @@ class MY_Model extends CI_Model
      */
     public function get_all()
     {
-        $this->_run_before_callbacks('get');
+        $this->trigger('before_get');
+
+        if ($this->soft_delete && $this->_temporary_with_deleted !== TRUE)
+        {
+            $this->db->where($this->soft_delete_key, FALSE);
+        }
 
         $result = $this->db->get($this->_table)
                            ->{$this->_return_type(1)}();
@@ -152,7 +210,7 @@ class MY_Model extends CI_Model
 
         foreach ($result as &$row)
         {
-            $row = $this->_run_after_callbacks('get', array( $row ));
+            $row = $this->trigger('after_get', $row);
         }
 
         return $result;
@@ -173,15 +231,15 @@ class MY_Model extends CI_Model
 
         if ($valid)
         {
-            $data = $this->_run_before_callbacks('create', array( $data ));
+            $data = $this->trigger('before_create', $data);
 
             $this->db->insert($this->_table, $data);
             $insert_id = $this->db->insert_id();
 
-            $this->_run_after_callbacks('create', array( $data, $insert_id ));
-            
+            $this->trigger('after_create', $insert_id);
+
             return $insert_id;
-        } 
+        }
         else
         {
             return FALSE;
@@ -210,7 +268,7 @@ class MY_Model extends CI_Model
     {
         $valid = TRUE;
 
-        $data = $this->_run_before_callbacks('update', array( $data, $primary_value ));
+        $data = $this->trigger('before_update', $data);
 
         if ($skip_validation === FALSE)
         {
@@ -222,7 +280,8 @@ class MY_Model extends CI_Model
             $result = $this->db->where($this->primary_key, $primary_value)
                                ->set($data)
                                ->update($this->_table);
-            $this->_run_after_callbacks('update', array( $data, $primary_value, $result ));
+
+            $this->trigger('after_update', array($data, $result));
 
             return $result;
         }
@@ -239,7 +298,7 @@ class MY_Model extends CI_Model
     {
         $valid = TRUE;
 
-        $data = $this->_run_before_callbacks('update', array( $data, $primary_values ));
+        $data = $this->trigger('before_update', $data);
 
         if ($skip_validation === FALSE)
         {
@@ -251,7 +310,8 @@ class MY_Model extends CI_Model
             $result = $this->db->where_in($this->primary_key, $primary_values)
                                ->set($data)
                                ->update($this->_table);
-            $this->_run_after_callbacks('update', array( $data, $primary_values, $result ));
+
+            $this->trigger('after_update', array($data, $result));
 
             return $result;
         }
@@ -270,13 +330,13 @@ class MY_Model extends CI_Model
         $data = array_pop($args);
         $this->_set_where($args);
 
-        $data = $this->_run_before_callbacks('update', array( $data, $args ));
+        $data = $this->trigger('before_update', $data);
 
         if ($this->_run_validation($data))
         {
             $result = $this->db->set($data)
                                ->update($this->_table);
-            $this->_run_after_callbacks('update', array( $data, $args, $result ));
+            $this->trigger('after_update', array($data, $result));
 
             return $result;
         }
@@ -291,10 +351,10 @@ class MY_Model extends CI_Model
      */
     public function update_all($data)
     {
-        $data = $this->_run_before_callbacks('update', array( $data ));
+        $data = $this->trigger('before_update', $data);
         $result = $this->db->set($data)
                            ->update($this->_table);
-        $this->_run_after_callbacks('update', array( $data, $result ));
+        $this->trigger('after_update', array($data, $result));
 
         return $result;
     }
@@ -304,10 +364,20 @@ class MY_Model extends CI_Model
      */
     public function delete($id)
     {
-        $data = $this->_run_before_callbacks('delete', array( $id ));
-        $result = $this->db->where($this->primary_key, $id)
-                           ->delete($this->_table);
-        $this->_run_after_callbacks('delete', array( $id, $result ));
+        $this->trigger('before_delete', $id);
+
+        $this->db->where($this->primary_key, $id);
+
+        if ($this->soft_delete)
+        {
+            $result = $this->db->update($this->_table, array( $this->soft_delete_key => TRUE ));
+        }
+        else
+        {
+            $result = $this->db->delete($this->_table);
+        }
+
+        $this->trigger('after_delete', $result);
 
         return $result;
     }
@@ -320,9 +390,18 @@ class MY_Model extends CI_Model
         $where = func_get_args();
         $this->_set_where($where);
 
-        $data = $this->_run_before_callbacks('delete', array( $where ));
-        $result = $this->db->delete($this->_table);
-        $this->_run_after_callbacks('delete', array( $where, $result ));
+        $where = $this->trigger('before_delete', $where);
+
+        if ($this->soft_delete)
+        {
+            $result = $this->db->update($this->_table, array( $this->soft_delete_key => TRUE ));
+        }
+        else
+        {
+            $result = $this->db->delete($this->_table);
+        }
+
+        $this->trigger('after_delete', $result);
 
         return $result;
     }
@@ -332,19 +411,96 @@ class MY_Model extends CI_Model
      */
     public function delete_many($primary_values)
     {
-        $data = $this->_run_before_callbacks('delete', array( $primary_values ));
-        $result = $this->db->where_in($this->primary_key, $primary_values)
-                           ->delete($this->_table);
-        $this->_run_after_callbacks('delete', array( $primary_values, $result ));
+        $primary_values = $this->trigger('before_delete', $primary_values);
+
+        $this->db->where_in($this->primary_key, $primary_values);
+
+        if ($this->soft_delete)
+        {
+            $result = $this->db->update($this->_table, array( $this->soft_delete_key => TRUE ));
+        }
+        else
+        {
+            $result = $this->db->delete($this->_table);
+        }
+
+        $this->trigger('after_delete', $result);
 
         return $result;
     }
 
-    public function deleted($column = NULL , $status = FALSE)
+    /* --------------------------------------------------------------
+     * RELATIONSHIPS
+     * ------------------------------------------------------------ */
+
+    public function with($relationship)
     {
-        $column = (is_null($column)) ? $this->_table.'.deleted' : $column ;
-        $this->db->where($column, $status);
+        $this->_with[] = $relationship;
+
+        if (!in_array('relate', $this->after_get))
+        {
+            $this->after_get[] = 'relate';
+        }
+
         return $this;
+    }
+
+    public function join($table, $primary_key ='id', $join = 'left')
+    {
+        $columns = self::_attr_selected($table);
+        $foreign_key = singular($table) . '_id';
+
+        $this->db->select($this->_table.'.*');
+        $this->db->select($columns);
+        $this->db->join($table, $table.'.'.$primary_key. '=' .$this->_table.'.'.$foreign_key, $join);
+        return $this;
+    }
+
+    public function relate($row)
+    {
+        foreach ($this->belongs_to as $key => $value)
+        {
+            if (is_string($value))
+            {
+                $relationship = $value;
+                $options = array( 'primary_key' => $value . '_id', 'model' => $value . '_model' );
+            }
+            else
+            {
+                $relationship = $key;
+                $options = $value;
+            }
+
+            if (in_array($relationship, $this->_with))
+            {
+                $this->load->model($options['model']);
+                $row->{$relationship} = $this->{$options['model']}->get($row->{$options['primary_key']});
+            }
+        }
+
+        foreach ($this->has_many as $key => $value)
+        {
+            if (is_string($value))
+            {
+                $relationship = $value;
+                $options = array( 'primary_key' => singular($this->_table) . '_id', 'model' => singular($value) . '_model' );
+            }
+            else
+            {
+                $relationship = $key;
+                $options = $value;
+            }
+
+            if (in_array($relationship, $this->_with))
+            {
+                $this->load->model($options['model']);
+                $row->{$relationship} = $this->{$options['model']}->get_many_by($options['primary_key'], $row->{$this->primary_key});
+            }
+        }
+
+        // $this->_with = array();
+
+        return $row;
     }
 
     /* --------------------------------------------------------------
@@ -368,12 +524,12 @@ class MY_Model extends CI_Model
             $value = $args[0];
         }
 
-        $this->_run_before_callbacks('get', array( $key, $value ));
+        $this->trigger('before_dropdown', array( $key, $value ));
 
         $result = $this->db->select(array($key, $value))
+                           ->order_by($value)
                            ->get($this->_table)
                            ->result();
-        $this->_run_after_callbacks('get', array( $key, $value, $result ));
 
         $options = array();
 
@@ -381,7 +537,9 @@ class MY_Model extends CI_Model
         {
             $options[$row->{$key}] = $row->{$value};
         }
-        
+
+        $options = $this->trigger('after_dropdown', $options);
+
         return $options;
     }
 
@@ -462,6 +620,66 @@ class MY_Model extends CI_Model
         return $this;
     }
 
+    /**
+     * Don't care about soft deleted rows on the next call
+     */
+    public function with_deleted()
+    {
+        $this->_temporary_with_deleted = TRUE;
+        return $this;
+    }
+
+    /* --------------------------------------------------------------
+     * OBSERVERS
+     * ------------------------------------------------------------ */
+
+    /**
+     * MySQL DATETIME created_at and updated_at
+     */
+    public function created_at($row)
+    {
+        $row['created_at'] = date('Y-m-d H:i:s');
+        return $row;
+    }
+
+    public function updated_at($row)
+    {
+        $row['updated_at'] = date('Y-m-d H:i:s');
+        return $row;
+    }
+
+    /**
+     * Serialises data for you automatically, allowing you to pass
+     * through objects and let it handle the serialisation in the background
+     */
+    public function serialize($row)
+    {
+        foreach ($this->callback_parameters as $column)
+        {
+            $row[$column] = serialize($row[$column]);
+        }
+
+        return $row;
+    }
+
+    public function unserialize($row)
+    {
+        foreach ($this->callback_parameters as $column)
+        {
+            if (is_array($row))
+            {
+                $row[$column] = unserialize($row[$column]);
+            }
+            else
+            {
+                $row->$column = unserialize($row->$column);
+            }
+        }
+
+        return $row;
+    }
+
+
     /* --------------------------------------------------------------
      * QUERY BUILDER DIRECT ACCESS METHODS
      * ------------------------------------------------------------ */
@@ -499,39 +717,23 @@ class MY_Model extends CI_Model
      * ------------------------------------------------------------ */
 
     /**
-     * Run the before_ callbacks, each callback taking a $data
-     * variable and returning it
+     * Trigger an event and call its observers
      */
-    private function _run_before_callbacks($type, $params = array())
+    public function trigger($event, $data = FALSE)
     {
-        $name = 'before_' . $type;
-        $data = (isset($params[0])) ? $params[0] : FALSE;
-
-        if (!empty($this->$name))
+        if (isset($this->$event) && is_array($this->$event))
         {
-            foreach ($this->$name as $method)
+            foreach ($this->$event as $method)
             {
-                $data += call_user_func_array(array($this, $method), $params);
-            }
-        }
+                if (strpos($method, '('))
+                {
+                    preg_match('/([a-zA-Z0-9\_\-]+)(\(([a-zA-Z0-9\_\-\., ]+)\))?/', $method, $matches);
 
-        return $data;
-    }
+                    $method = $matches[1];
+                    $this->callback_parameters = explode(',', $matches[3]);
+                }
 
-    /**
-     * Run the after_ callbacks, each callback taking a $data
-     * variable and returning it
-     */
-    private function _run_after_callbacks($type, $params = array())
-    {
-        $name = 'after_' . $type;
-        $data = (isset($params[0])) ? $params[0] : FALSE;
-
-        if (!empty($this->$name))
-        {
-            foreach ($this->$name as $method)
-            {
-                $data = call_user_func_array(array($this, $method), $params);
+                $data = call_user_func_array(array($this, $method), array($data));
             }
         }
 
@@ -585,6 +787,23 @@ class MY_Model extends CI_Model
         }
     }
 
+    /* --------------------------------------------------------------
+     * MULTIPLE DB LOADING
+     * ------------------------------------------------------------ */
+
+    private function _set_database()
+    {
+        if (!$this->_db)
+        {
+            $this->load->database();
+        }
+        else
+        {
+            $this->db = $this->load->database($this->_db, TRUE);
+        }
+    }
+
+
     /**
      * Set WHERE parameters, cleverly
      */
@@ -607,5 +826,19 @@ class MY_Model extends CI_Model
     {
         $method = ($multi) ? 'result' : 'row';
         return $this->_temporary_return_type == 'array' ? $method . '_array' : $method;
+    }
+
+    /**
+     * Return columns of join table minus excluded attributes
+     */
+    private function _attr_selected($table)
+    {
+        $fields = $this->db->list_fields($table);
+
+        foreach ($this->excluded_attr as $key => $value) {
+            $column = array_search($value, $fields);
+            unset($fields[$column]);
+        }
+        return array_values($fields);
     }
 }
